🤖 Agentic AI/class_05/myApp/pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ArAhmadRaza", email = "arahmadraza5570@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.16.1",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.15",
]

[project.scripts]
myapp = "myapp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
